py/tools/releasekit/pyproject.toml
# Copyright 2025 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
#
# SPDX-License-Identifier: Apache-2.0

[project]
authors = [
  { name = "Google" },
  { name = "Yesudeep Mangalapilly", email = "yesudeep@google.com" },
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Software Development :: Build Tools",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Typing :: Typed",
  "Private :: Do Not Upload",
]
dependencies = [
  "aiofiles>=24.1.0",                         # Non-blocking file I/O for async Workspace protocol
  "packaging>=24.0",                          # PEP 508 dependency specifier parsing
  "tomlkit>=0.13.0",                          # Style-preserving TOML manipulation
  "structlog>=25.1.0",                        # Structured logging
  "rich>=13.0.0",                             # Rich terminal UI + progress
  "rich-argparse>=1.6.0",                     # Colorful CLI help
  "argcomplete>=3.0.0",                       # Shell tab completion
  "jinja2>=3.1.0",                            # Release notes templates
  "PyJWT[crypto]>=2.8.0",                     # JWT signature verification (JWKS/RS256/ES256)
  "diagnostic>=3.0.0",                        # Rust-style error rendering
  "httpx>=0.27.0",                            # Async HTTP with connection pooling (PyPIBackend)
  "opentelemetry-api>=1.20.0",                # Distributed tracing API
  "opentelemetry-sdk>=1.20.0",                # Distributed tracing SDK
  "pypi-attestations>=0.0.20",                # PEP 740 attestation generation + verification
  "sigstore>=3.0.0",                          # Keyless artifact signing via Sigstore OIDC
  "strenum>=0.4.15; python_version < '3.11'", # StrEnum backport for 3.10
]
description = "Release orchestration for polyglot monorepos"
license = "Apache-2.0"
name = "releasekit"
readme = "README.md"
requires-python = ">=3.10"
version = "0.1.0"

[project.scripts]
releasekit = "releasekit.cli:_main"

[build-system]
build-backend = "hatchling.build"
requires      = ["hatchling"]

[tool.hatch.build.targets.wheel]
packages = ["src/releasekit"]

[tool.hatch.build.targets.wheel.force-include]
"LICENSE" = "releasekit/LICENSE"

[dependency-groups]
dev = [
  "jsonschema>=4.26.0",
  "pyyaml>=6.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.25.0",
  "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
addopts    = "--cov=releasekit --cov-config=pyproject.toml --cov-report=term-missing"
pythonpath = ["."]
testpaths  = ["tests"]

[tool.coverage.run]
branch = true
omit   = ["src/releasekit/cli.py", "src/releasekit/ui.py"]
source = ["releasekit"]

[tool.coverage.report]
exclude_also = [
  "if __name__ == .__main__.",
  "if TYPE_CHECKING:",
  "pragma: no cover",
  "@overload",
  "raise NotImplementedError",
]
fail_under = 85
show_missing = true
skip_covered = false
